Enabled froala_editor, email_notifications, delete_permissions etc

+++ blog/views.py
@@ -3,13 +3,19 @@ from django.http import HttpResponse, HttpResponseRedirect, JsonResponse
 
 from .models import Post
 from .forms import PostForm
+from .notify import (
+    send_post_create_notification,
+    got_hit_notification
+)
 
 
 def show_posts(request):
-    return render(request, 'blog/show_posts.html', {'object_list': Post.objects.all().order_by('last_edit_date'), 'all': True})
+    # got_hit_notification(page="blog/show_posts")
+    return render(request, 'blog/show_posts.html', {'object_list': Post.objects.all().order_by('-last_edit_date'), 'all': True})
 
 
 def post_detail(request, pk):
+    # got_hit_notification(page="blog/post_detail")
     return render(request, 'blog/post_detail.html', {'post': Post.objects.filter(id=pk)[0]})
 
 
@@ -18,6 +24,7 @@ def post_create(request):
         form = PostForm(request.POST or None)
         if form.is_valid():
             instance = form.save(commit=False)
+            # send_post_create_notification(form);
             instance.save()
             # return HttpResponseRedirect(instance.get_absolute_url())
     form = PostForm()
@@ -41,6 +48,14 @@ def post_edit(request, pk=None):
     return render(request, 'blog/post_create.html', context)
 
 def post_delete(request, pk=None):
-    instance = get_object_or_404(Post, id=pk)
-    instance.delete();
-    return JsonResponse({'message': "Successfully deleted!", 'status': 'OK'})
+    user = request.user
+    message = '<div class="alert alert-%s alert-dismissible" role="alert"><strong>%s!</strong> %s</div>'
+
+    if user.is_authenticated and user.is_superuser:
+        instance = get_object_or_404(Post, id=pk)
+        instance.delete();
+        message = message % ('success', "Success", 'Successfully deleted!')
+        return JsonResponse({'message': message, 'status': 'deleted'})
+    else:
+        message = message % ('danger', "Error", "You can't delete that post!")
+        return JsonResponse({'message': message, 'status': 'not deleted'})
